added subscribe.tsx to the footer
