--- api/subscribers.csv
@@ -1,1 +1,0 @@
-Email
