+++ package.json
@@ -13,12 +13,14 @@
     "next": "14.2.3",
     "react": "^18",
     "react-dom": "^18",
-    "react-icons": "^5.2.1"
+    "react-icons": "^5.2.1",
+    "validator": "^13.12.0"
   },
   "devDependencies": {
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/validator": "^13.12.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
     "postcss": "^8",
